use u16 for squares

+++ chess_foundation/src/board_helper.rs
@@ -1,10 +1,10 @@
 use crate::coord::Coord;
 
-pub fn file_index(square_index: i32) -> i32 {
+pub fn file_index(square_index: u16) -> u16 {
     square_index & 0b000111
 }
 
-pub fn rank_index(square_index: i32) -> i32 {
+pub fn rank_index(square_index: u16) -> u16 {
     square_index >> 3
 }
 
@@ -17,6 +17,6 @@ pub fn square_index_to_board_row_col(index: i32) -> (usize, usize) {
     let row = 7 - index / 8;
     (row as usize, col as usize)
 }
-pub fn board_row_col_to_square_index(row: usize, col: usize) -> i32 {
-    ((7 - row) * 8 + col) as i32
+pub fn board_row_col_to_square_index(row: usize, col: usize) -> u16 {
+    ((7 - row) * 8 + col) as u16
 }
